ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2
google-genai>=0.2.2
python-dotenv>=1.0.1
